little modifications
I cleaned a little bit of the unused imports and methods of the Budget module

I removed some unused buttons of the .forms from the Budget module

I modified the code that creates the PDFs. Now when you create a PDF of a Budget, if the file exists, it creates a copy.

little more fixes blabla